// src/app/models/entreprise/demandeProfil.ts
import { EntrepriseAccount } from "./entreprise";

export class DemandeProfil {
  oid: number;
  poste: string;
  experience: string;
  domaineFormaton: string;
  dateline: string;
  nivScolaire: string;
  competence: string;
  civiliy: string;
  statutMatrimonial: string;
  nbProfil: string;
  duree: string;
  naisCancelme: string;
  status: string;
  typeContrat: string;
  horaireTravail: string;
  recrutement: string;
  lieuTravail: string;
  description: string;
  created_at: string;
  updated_at: string;
  entrepriseAccount: EntrepriseAccount;
}